Make restaurant names unique
This is a tiny commit mostly to write this note. Drizzle _REALLY_ does
not play nicely with circular references ('restaurants.currentMenu` ->
'menus.id' && 'menus.restaurantId' -> 'restaurants.id'). Because of
this, a lot of the runtime reflection breaks and as a consequence,
'drizzle-seed' throws an error when 'isUnique: true' is passed to the
restaurants refinement ('src/app/api/demo/seed/route.ts:29'). _HOWEVER_,
it still sucessfully seeds the database so I'm going to leave it as is
(even with the error).

I also bumped 'drizzle-kit' to '^0.30.5' because 'pnpm run db:push' was
failing with 'ELIFECYCLE'.